% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_grapher.R
\name{grapher_labels}
\alias{grapher_labels}
\title{Add labels to the grapher}
\usage{
grapher_labels(grapher, title = "", subtitle = "", note = "", source = "")
}
\arguments{
\item{grapher}{An object of class grapher}

\item{title}{The title of the graph}

\item{subtitle}{The subtitle displayed underneath the title}

\item{note}{A footnote for the graph}

\item{source}{The source of data}
}
\description{
Add labels to the grapher
}
