% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_grapher.R
\name{grapher_line}
\alias{grapher_line}
\title{Add a line graph to the grapher}
\usage{
grapher_line(
  grapher,
  selected = c("United Kingdom", "France", "Spain", "Ireland")
)
}
\arguments{
\item{grapher}{An object of class "grapher".}

\item{selected}{The entities displayed when the graph first loads.}
}
\value{

}
\description{
Add a line graph to the grapher
}
