% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_grapher.R
\name{grapher_map}
\alias{grapher_map}
\title{Add a map to the grapher}
\usage{
grapher_map(grapher, palette = "Greens", bins = "auto")
}
\arguments{
\item{grapher}{An object of class "grapher".}

\item{palette}{An RColorBrewer palette.}

\item{bins}{The steps in the map legend bins.}
}
\description{
Add a map to the grapher
}
