% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid.R
\name{owid}
\alias{owid}
\title{Get data from Our World in Data}
\usage{
owid(datasets, id = NULL, ...)
}
\arguments{
\item{datasets}{A dataframe returned by get_owid_datasets().}

\item{id}{The id of a dataset.}

\item{...}{Further arguments passed on to read_csv.}
}
\value{
A tibble of an owid dataset with the added class 'owid'.
}
\description{
Get OWID datasets from the OWID datasets github repo.
}
\examples{
ds <- owid_get_datasets()
owid_search(ds, "meat")
id <- owid_search(ds, "Meat consumption in EU28") \%>\% pull(id)
meat <- owid(ds, id)
}
