% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid.R
\name{owid_covid}
\alias{owid_covid}
\title{Get the Our World in Data covid-19 dataset}
\usage{
owid_covid()
}
\value{
A dataframe with multiple variables on the covid-19 pandemic.
}
\description{
Get the Our World in Data covid-19 dataset
}
