% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_url.R
\name{owid_description}
\alias{owid_description}
\title{Show the description of the data}
\usage{
owid_description(data)
}
\arguments{
\item{data}{A dataframe returned from owid() of class "owid".}
}
\value{

}
\description{
Show the description of the data
}
