% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_get_datasets.R
\name{owid_get_datasets}
\alias{owid_get_datasets}
\title{Get a dataframe of the titles of all owid datasets}
\usage{
owid_get_datasets()
}
\value{

}
\description{
This function web scrapes the OWID datasets github repo to create a dataframe that lists all of the available datasets and gives each one a unique value.
}
