% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_map.R
\name{owid_map}
\alias{owid_map}
\title{Title
Create a world map chloropleth of owid data.}
\usage{
owid_map(data, col = 3, palette = "Reds", mode = "plot")
}
\arguments{
\item{data}{A tibble returned by `owid()`. This tibble must have country names as the Entity, not all data returned by `owid()` will be like this.}

\item{col}{The column number to be treated as the value.}

\item{palette}{The `RColorBrewer` palette to be used.}
}
\value{
Either a ggplot
}
\description{
A function to easily create a world map chloropleth of owid data.
}
\examples{
`owid_map()`
}
