% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_plot.R
\name{owid_plot}
\alias{owid_plot}
\title{Plot an owid dataset}
\usage{
owid_plot(
  data = NULL,
  col = 3,
  summarise = TRUE,
  filter = NULL,
  years = NULL,
  show.all = FALSE
)
}
\arguments{
\item{data}{A tibble returned from `owid()`}

\item{col}{The column number of the value to be plotted. Defaults to 3, and must be at least 3.}

\item{summarise}{If TRUE, plot takes the mean value. If FALSE, each Entity is plotted, it is recommended to use this in conjunction with the filter argument to avoid too many Entity's being plotted.}

\item{filter}{The Entity's to include in the plot.}

\item{years}{The Years to be included in the plot.}
}
\value{
A ggplot object.
}
\description{
A wrapper around ggplot to provide an quick visualisation of owid data.
}
\examples{
ds <- owid_get_datasets()
owid_search(ds, "meat")
id <- owid_search(ds, "Meat consumption in EU28") \%>\% pull(id)
meat <- owid(ds, id)

owid_plot(meat)
}
