% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_readme.R
\name{owid_readme}
\alias{owid_readme}
\title{Print the readme for OWID data.}
\usage{
owid_readme(data)
}
\arguments{
\item{data}{An owid dataset.}
}
\value{

}
\description{
Prints the readme for OWID data as found on the OWID datasets github repo.
}
\examples{
ds <- owid_get_datasets()
id <- owid_search(ds, "YouGov-Imperial COVID-19 Behavior Tracker")$id

covid_behav <- owid(id, ds)
owid_readme(covid_behav)
}
