% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid.R
\name{owid_search}
\alias{owid_search}
\title{Search the data sources used in OWID charts}
\usage{
owid_search(term)
}
\arguments{
\item{term}{A search term}
}
\value{

}
\description{
Search the data sources used in OWID charts
}
