% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{owid_source}
\alias{owid_source}
\title{Get source information on an OWID dataset}
\usage{
owid_source(data)
}
\arguments{
\item{data}{A tibble returned from owid().}
}
\value{
Displays the information in an easy to read format in the R console, also returns a list of data information.
}
\description{
A function to get source information from an OWID dataset and display it in the R console.
}
\examples{
\donttest{
rights <- owid("human-rights-scores")
owid_source(rights)
}
}
