% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_url.R
\name{owid_sources}
\alias{owid_sources}
\title{Show the source information of an owid dataset}
\usage{
owid_sources(data)
}
\arguments{
\item{data}{A dataframe returned from owid() of class "owid".}
}
\value{

}
\description{
Show the source information of an owid dataset
}
