% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owidR-package.R
\docType{package}
\name{owidR-package}
\alias{owidR}
\alias{owidR-package}
\title{owidR: Import Data from Our World in Data}
\description{
Search and download the data used in over 3,000 Our World in Data charts (ourworldindata.org).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/piersyork/owidR}
  \item \url{https://github.com/owid/covid-19-data/}
  \item \url{https://ourworldindata.org/}
}

}
\keyword{internal}
