% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_owid.R
\name{pal_owid}
\alias{pal_owid}
\title{Colour palettes based on the colours used by Our World in Data}
\usage{
pal_owid(alpha)
}
\arguments{
\item{alpha}{Transparency level, a real number in (0, 1).}
}
\description{
Colour palettes based on the colours used by Our World in Data
}
