% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.grapher.R
\name{print.grapher}
\alias{print.grapher}
\title{The print method for owid_grapher()}
\usage{
\method{print}{grapher}(x, ...)
}
\arguments{
\item{x}{An object of class "grapher"}

\item{...}{Further arguments for future versions}
}
\description{
Create the final graph and present it in the viewer
}
