% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_owid.R
\name{scale_fill_owid}
\alias{scale_fill_owid}
\alias{scale_colour_owid}
\alias{scale_color_owid}
\title{Our World in Data Colour Scales}
\usage{
scale_fill_owid(alpha = 1, ...)

scale_colour_owid(alpha = 1, ...)

scale_color_owid(alpha = 1, ...)
}
\arguments{
\item{alpha}{Transparency level, a real number in (0, 1).}

\item{...}{additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\value{
A ggproto object to be used in the context of ggplot2.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions have been deprecated to simplify the owidR package, you are encouraged to use ggplot2 instead: https://ggplot2-book.org
}
