% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_chart}
\alias{view_chart}
\title{View an OWID chart in your browser}
\usage{
view_chart(x)
}
\arguments{
\item{x}{Either a dataset returned by owid(), or a chart_id.}
}
\value{
Opens the chart in your browser.
}
\description{
A function that opens the original OWID chart in your browser.
}
\examples{
\donttest{
firearm_suicide <- owid("suicide-rate-by-firearm")
view_chart(firearm_suicide)
}
}
