% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_map.R
\name{world_map_data}
\alias{world_map_data}
\title{Get world map data.}
\usage{
world_map_data()
}
\value{
An object of class sf.
}
\description{
Function that returns a simple feature collection of class sf. Map data is from naturalearthdata.com. Designed to be used internally.
}
