% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid.R
\name{owid}
\alias{owid}
\title{Get data from Our World in Data}
\usage{
owid(chart_id = NULL, rename = NULL, tidy.date = TRUE, ...)
}
\arguments{
\item{chart_id}{The chart_id as returned by owid_search}

\item{rename}{Rename the value column. Currently only works if their is just one value column.}

\item{tidy.date}{If TRUE then a year column that should be a date column will automatically detected and transformed. If FALSE then the Year column will be kept as is. Defaults to TRUE.}

\item{...}{Not to be used.}
}
\value{
A tibble of an owid dataset with the added class 'owid'.
}
\description{
Get a dataset used in an OWID chart.
}
\examples{
\donttest{
owid_search("emissions")
emissions <- owid("per-capita-ghg-emissions")
}

}
