% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_map.R
\name{owid_map}
\alias{owid_map}
\title{Title
Create a world map chloropleth of owid data.}
\usage{
owid_map(data, col = 3, palette = "Reds", mode = "plot")
}
\arguments{
\item{data}{A tibble returned by `owid()`. This tibble must have country names as the Entity, not all data returned by `owid()` will be like this.}

\item{col}{The column number to be treated as the value.}

\item{palette}{The `RColorBrewer` palette to be used.}

\item{mode}{If `"plot"`, the output will be a ggplot2 map. If `"view"`, the output will be a leaflet interactive map.}
}
\value{
Either a ggplot2
}
\description{
A function to easily create a world map chloropleth of owid data.
}
\examples{
`ds <- owid_get_datasets()`
`id <- owid_search(ds, "Mental and substance use disorder")$id`
`mental <- owid(ds, id)`

# simple ggplot2 map
`owid_map(mental)`

# interavtive map with blue palette
`owid_map(mental, mode = "view", palette = "Blues")`

}
