% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_search.R
\name{owid_search}
\alias{owid_search}
\title{Search for terms in dataframe of OWID datasets}
\usage{
owid_search(datasets, term = NULL)
}
\arguments{
\item{datasets}{A dataframe returned by owid_get_datasets().}

\item{term}{A search term. owid_search() uses grep with case ignored, so either a key word or a regular expression is acceptable.}
}
\value{
A tibble with the dataset title's matching the search.
}
\description{
Search through dataframe of OWID datasets using key word or regular expression.
}
\examples{
ds <- owid_get_datasets()
owid_search(ds, "meat")
# Currently id is generated from the rownumber, as such id's are not fixed and future
# data uploads will change the id. It is recommended to have a specific search and
# then pull the id from that search.
id <- owid_search(ds, "Meat consumption in EU28") \%>\% pull(id)
meat <- owid(ds, id)
}
