% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid.R
\name{owid_search}
\alias{owid_search}
\title{Search the data sources used in OWID charts}
\usage{
owid_search(term)
}
\arguments{
\item{term}{A search term}
}
\value{
A matrix of chart titles and chard ids
}
\description{
Search the data sources used in OWID charts
}
\examples{
\donttest{
# returns the titles and chart_ids of all charts containing the word 'emissions'
owid_search("emissions")
}

}
