% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_url.R
\name{owid_url}
\alias{owid_url}
\title{Get the GitHub url for the data}
\usage{
owid_url(data)
}
\arguments{
\item{data}{A tibble returned from `owid()`}
}
\value{
The GitHub url as a character string.
}
\description{
Get the GitHub url for the data
}
