% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_url.R
\name{owid_url}
\alias{owid_url}
\title{Get the GitHub and source URLs for a dataset}
\usage{
owid_url(data)
}
\arguments{
\item{data}{A tibble returned from `owid()`}
}
\value{
Returns a named vector of the owid GitHub URL and the source URL.
}
\description{
Get the GitHub and source URLs for a dataset
}
