% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owidR-package.R
\docType{package}
\name{owidR-package}
\alias{owidR}
\alias{owidR-package}
\title{owidR: A Package for Importing Data from Our World in Data}
\description{
Imports data from the Our World in Data website, offering easy to use functions for searching for datasets, downloading them into R and visualising them.
}
\seealso{
Useful links:
\itemize{
  \item \url{<https://github.com/piersyork/owidR>}
  \item \url{<https://github.com/owid/covid-19-data/>}
  \item \url{<https://ourworldindata.org/>}
}

}
\keyword{internal}
